move utils/Autocomplete to shared/Autocomplete
